Prevent ticks from being processed multiple times
When the daemon checks in to the server multiple times before a tick
has run it can pull the same data repeatedly, as the console command to
erase the stats data in memory has not gone through yet. This change
records the last processed tick and refuses to process anything before
that.